[Fix] Pattern Editor: Reverted playback cursor display behaviour to how it worked before OpenMPT 1.22 [Fix] FT2 Compatibility: Fixed Rxy a bit more (http://forum.openmpt.org/index.php?topic=4999.0) [Mod] OpenMPT: Version is now 1.22.02.01
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@1923 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/MainFrm.cpp
@@ -1637,7 +1637,7 @@ void CMainFrame::InitPreview()
 	// Avoid global volume ramping when trying samples in the treeview.
 	m_WaveFile.m_nDefaultGlobalVolume = m_WaveFile.m_nGlobalVolume = MAX_GLOBAL_VOLUME;
 	m_WaveFile.SetMixLevels(mixLevels_117RC3);
-	m_WaveFile.m_nSamplePreAmp = m_WaveFile.GetPlayConfig().getNormalSamplePreAmp();
+	m_WaveFile.m_nSamplePreAmp = static_cast<uint32>(m_WaveFile.GetPlayConfig().getNormalSamplePreAmp());
 	m_WaveFile.m_nDefaultTempo = 125;
 	m_WaveFile.m_nDefaultSpeed = 6;
 	m_WaveFile.m_nType = MOD_TYPE_MPT;
